d_a_npc_uk: getMsg 93.47->100% (remove bogus clrFlag(0x8) - retail has none in getMsg; getFirstTalk u32->u16 so retail's mr r4,r3 full-register pass matches, no clrlwi); .o 01:18

+++ include/d/actor/d_a_npc_uk.h
@@ -104,7 +104,7 @@ public:
     int getShapeType();
     u16 getCaughtFlag();
     u16 getFoundFlag();
-    u32 getFirstTalk();
+    u16 getFirstTalk();
     BOOL chkGameStart();
     BOOL chkPositioning(f32, f32, f32, s16, s16);
     u8 nextVisitMode();
